fix negative eut
