+++ src/main/java/gregtech/api/capability/impl/RecipeLogicEnergy.java
@@ -25,33 +25,8 @@ public class RecipeLogicEnergy extends AbstractRecipeLogic {
     }
 
     @Override
-    protected long getEnergyInputPerSecond() {
-        return energyContainer.get().getInputPerSec();
-    }
-
-    @Override
-    protected long getEnergyStored() {
-        return energyContainer.get().getEnergyStored();
-    }
-
-    @Override
-    protected long getEnergyCapacity() {
-        return energyContainer.get().getEnergyCapacity();
-    }
-
-    @Override
-    protected boolean drawEnergy(long recipeEUt, boolean simulate) {
-        long resultEnergy = getEnergyStored() - recipeEUt;
-        if (resultEnergy >= 0L && resultEnergy <= getEnergyCapacity()) {
-            if (!simulate) energyContainer.get().changeEnergy(-recipeEUt);
-            return true;
-        }
-        return false;
-    }
-
-    @Override
-    public long getMaxVoltage() {
-        return Math.max(energyContainer.get().getInputVoltage(), energyContainer.get().getOutputVoltage());
+    protected IEnergyContainer getEnergyContainer() {
+        return energyContainer.get();
     }
 
     @Override
