+++ src/main/java/gregtech/api/capability/impl/SteamMultiblockRecipeLogic.java
@@ -1,6 +1,7 @@
 package gregtech.api.capability.impl;
 
 import gregtech.api.GTValues;
+import gregtech.api.capability.IEnergyContainer;
 import gregtech.api.capability.IMultipleTankHandler;
 import gregtech.api.metatileentity.multiblock.RecipeMapSteamMultiblockController;
 import gregtech.api.recipes.Recipe;
@@ -45,6 +46,11 @@ public class SteamMultiblockRecipeLogic extends AbstractRecipeLogic {
     public IFluidTank getSteamFluidTankCombined() {
         combineSteamTanks();
         return steamFluidTankCombined;
+    }
+
+    @Override
+    protected IEnergyContainer getEnergyContainer() {
+        return IEnergyContainer.DEFAULT;
     }
 
     @Override
